Move value semantics related tests to specific file
